+++ packages/server/src/routes/appRouter/iconHandler.test.ts
@@ -3,8 +3,11 @@ import { request, setTestApp } from 'axios-test-instance';
 import Koa from 'koa';
 
 import { appRouter } from '.';
-import { App } from '../../models';
+import { App, Organization } from '../../models';
 import * as appUtils from '../../utils/app';
+import { createTestSchema } from '../../utils/test/testSchema';
+
+beforeAll(createTestSchema('iconHandler'));
 
 beforeAll(async () => {
   request.defaults.responseType = 'arraybuffer';
@@ -13,8 +16,8 @@ beforeAll(async () => {
 
 it('should scale and serve the app icon', async () => {
   jest.spyOn(appUtils, 'getApp').mockResolvedValue({
-    icon: await readFixture('tux.png'),
-  } as Partial<App> as App);
+    app: new App({ icon: await readFixture('tux.png') }),
+  });
   const response = await request.get('/icon-150.png');
   expect(response.headers['content-type']).toBe('image/png');
   expect(response.data).toMatchImageSnapshot();
@@ -22,9 +25,11 @@ it('should scale and serve the app icon', async () => {
 
 it('should use the splash color if an opaque icon is requested', async () => {
   jest.spyOn(appUtils, 'getApp').mockResolvedValue({
-    definition: { theme: { splashColor: '#ff0000', themeColor: '#00ff00' } },
-    icon: await readFixture('tux.png'),
-  } as Partial<App> as App);
+    app: new App({
+      definition: { theme: { splashColor: '#ff0000', themeColor: '#00ff00' } },
+      icon: await readFixture('tux.png'),
+    }),
+  });
   const response = await request.get('/icon-52.png?opaque');
   expect(response.headers['content-type']).toBe('image/png');
   expect(response.data).toMatchImageSnapshot();
@@ -32,9 +37,11 @@ it('should use the splash color if an opaque icon is requested', async () => {
 
 it('should fall back to the theme color if splash color is undefined', async () => {
   jest.spyOn(appUtils, 'getApp').mockResolvedValue({
-    definition: { theme: { themeColor: '#00ff00' } },
-    icon: await readFixture('tux.png'),
-  } as Partial<App> as App);
+    app: new App({
+      definition: { theme: { themeColor: '#00ff00' } },
+      icon: await readFixture('tux.png'),
+    }),
+  });
   const response = await request.get('/icon-85.png?opaque');
   expect(response.headers['content-type']).toBe('image/png');
   expect(response.data).toMatchImageSnapshot();
@@ -42,9 +49,11 @@ it('should fall back to the theme color if splash color is undefined', async () 
 
 it('should fall back to a white background if neither theme color not splash color is defined', async () => {
   jest.spyOn(appUtils, 'getApp').mockResolvedValue({
-    definition: { theme: {} },
-    icon: await readFixture('tux.png'),
-  } as Partial<App> as App);
+    app: new App({
+      definition: { theme: {} },
+      icon: await readFixture('tux.png'),
+    }),
+  });
   const response = await request.get('/icon-24.png?maskable=true');
   expect(response.headers['content-type']).toBe('image/png');
   expect(response.data).toMatchImageSnapshot();
@@ -52,27 +61,31 @@ it('should fall back to a white background if neither theme color not splash col
 
 it('should fall back to a white background if theme is undefined', async () => {
   jest.spyOn(appUtils, 'getApp').mockResolvedValue({
-    definition: {},
-    icon: await readFixture('tux.png'),
-  } as Partial<App> as App);
+    app: new App({
+      definition: {},
+      icon: await readFixture('tux.png'),
+    }),
+  });
   const response = await request.get('/icon-235.png?maskable=true');
   expect(response.headers['content-type']).toBe('image/png');
   expect(response.data).toMatchImageSnapshot();
 });
 
 it('should fall back to the organization icon if no app app icon is defined', async () => {
-  jest.spyOn(appUtils, 'getApp').mockResolvedValue({
-    Organization: {
-      icon: await readFixture('nodejs-logo.png'),
-    },
-  } as Partial<App> as App);
+  const app = new App();
+  app.Organization = new Organization({
+    icon: await readFixture('nodejs-logo.png'),
+  });
+  jest.spyOn(appUtils, 'getApp').mockResolvedValue({ app });
   const response = await request.get('/icon-42.png');
   expect(response.headers['content-type']).toBe('image/png');
   expect(response.data).toMatchImageSnapshot();
 });
 
 it('should fall back to the Appsemble icon if no app or organization icon is defined', async () => {
-  jest.spyOn(appUtils, 'getApp').mockResolvedValue({ Organization: {} } as Partial<App> as App);
+  const app = new App();
+  app.Organization = new Organization();
+  jest.spyOn(appUtils, 'getApp').mockResolvedValue({ app });
   const response = await request.get('/icon-42.png');
   expect(response.headers['content-type']).toBe('image/png');
   expect(response.data).toMatchImageSnapshot();