+++ packages/server/src/utils/model.ts
@@ -15,7 +15,6 @@ import { argv } from './argv';
  *
  * @param record - The sequelize App model to normalize.
  * @param omittedValues - A list of fields to omit from the result.
- *
  * @returns An app resource that can be safely returned from the API.
  */
 export function getAppFromRecord(