+++ packages/studio/src/pages/apps/app/resources/resource/resource-definition-details/Endpoint/messages.ts
@@ -1,0 +1,12 @@
+import { defineMessages } from 'react-intl';
+
+export const messages = defineMessages({
+  get: 'Get a single {resourceName} resource',
+  query: 'Get all {resourceName} resources',
+  create: 'Create a new {resourceName} resource',
+  delete: 'Delete an existing {resourceName} resource',
+  $count: 'Count {resourceName} resources',
+  update: 'Update an existing {resourceName} resource',
+  role: 'Role: {role}',
+  public: 'Public',
+});